Dagger normalization procedure
The class BaseApplication was created to assure the initialization of the components required by the project injection support. Additionally the component was declared in the MainActivity context in order to allow the injection of the modules consumed by the dummy application;
The gradle file was modified to include the latest Dagger release;
The Android Manifest was modified in order to include the BaseApplication class, used to initialize the components implied by the Dagger configuration policy.